requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
passlib[bcrypt]
python-jose      # for JWT (or pyjwt)
python-multipart # for OAuth2 form parsing
alembic
